pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eden-emergent"
version = "1.0.0"
description = "EDEN v1.0 — Emergent Developmental Encoding Network"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "click>=8.1",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.13",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-cov>=4.1"]
vision = ["torchvision>=0.15"]
tui = ["textual>=0.47"]
neat = ["neat-python>=0.92"]
umap = ["umap-learn>=0.5", "scipy>=1.11"]
config = ["pyyaml>=6.0"]
profile = ["thop>=0.1.1"]

[project.scripts]
eden = "eden.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["eden*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
